add a reset --hard option (#46)
* add a reset --hard option

* move game file handling to a separate module

* handle both cases in the same function

* improve naming

* update changelog
